+++ client/jetbrains/src/main/java/com/sourcegraph/cody/CodyToolWindowContent.java
@@ -57,7 +57,6 @@ import com.sourcegraph.cody.recipes.FindCodeSmellsAction;
 import com.sourcegraph.cody.recipes.GenerateDocStringAction;
 import com.sourcegraph.cody.recipes.GenerateUnitTestAction;
 import com.sourcegraph.cody.recipes.ImproveVariableNamesAction;
-import com.sourcegraph.cody.recipes.OptimizeCodeAction;
 import com.sourcegraph.cody.recipes.RecipeRunner;
 import com.sourcegraph.cody.recipes.SummarizeRecentChangesRecipe;
 import com.sourcegraph.cody.recipes.TranslateToLanguageAction;
@@ -149,9 +148,6 @@ class CodyToolWindowContent implements UpdatableChat {
     // contextSearchButton.addActionListener(e -> recipeRunner.runContextSearch());
     // JButton releaseNotesButton = createWideButton("Generate release notes");
     // releaseNotesButton.addActionListener(e -> recipeRunner.runReleaseNotes());
-    JButton optimizeCodeButton = createRecipeButton("Optimize code");
-    optimizeCodeButton.addActionListener(
-        e -> new OptimizeCodeAction().executeRecipeWithPromptProvider(this, project));
     recipesPanel.add(explainCodeDetailedButton);
     recipesPanel.add(explainCodeHighLevelButton);
     recipesPanel.add(generateUnitTestButton);
@@ -163,7 +159,6 @@ class CodyToolWindowContent implements UpdatableChat {
     //    recipesPanel.add(fixupButton);
     //    recipesPanel.add(contextSearchButton);
     //    recipesPanel.add(releaseNotesButton);
-    recipesPanel.add(optimizeCodeButton);
 
     // Chat panel
     messagesPanel.setLayout(new VerticalFlowLayout(VerticalFlowLayout.TOP, 0, 0, true, true));